+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/EspError.scala
@@ -19,3 +19,5 @@ trait NotFoundError extends EspError
 trait BadRequestError extends EspError
 
 trait IllegalOperationError extends EspError
+
+trait ForbiddenError extends EspError