+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/CustomActionInvokerService.scala
@@ -1,11 +1,11 @@
 package pl.touk.nussknacker.ui.process.deployment
 
-import pl.touk.nussknacker.engine
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.ProcessIdWithName
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
-import pl.touk.nussknacker.restmodel.processdetails.BaseProcessDetails
+import pl.touk.nussknacker.ui.{EspError, NotFoundError}
 import pl.touk.nussknacker.ui.process.deployment.LoggedUserConversions.LoggedUserOps
+import pl.touk.nussknacker.ui.process.exception.ProcessIllegalAction
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository
 import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository.ProcessNotFoundError
 import pl.touk.nussknacker.ui.security.api.LoggedUser
@@ -17,7 +17,7 @@ trait CustomActionInvokerService {
   def invokeCustomAction(actionName: String, id: ProcessIdWithName, params: Map[String, String])(
       implicit loggedUser: LoggedUser,
       ec: ExecutionContext
-  ): Future[Either[CustomActionError, CustomActionResult]]
+  ): Future[CustomActionResult]
 
 }
 
@@ -30,41 +30,39 @@ class CustomActionInvokerServiceImpl(
     dispatcher: DeploymentManagerDispatcher,
     processStateService: ProcessStateService
 ) extends CustomActionInvokerService {
-  override def invokeCustomAction(actionName: String, id: ProcessIdWithName, params: Map[String, String])(
+  override def invokeCustomAction(
+      actionName: String,
+      id: ProcessIdWithName,
+      params: Map[String, String]
+  )(
       implicit user: LoggedUser,
       ec: ExecutionContext
-  ): Future[Either[CustomActionError, CustomActionResult]] = {
-
-    def createCustomAction(process: BaseProcessDetails[_]) =
-      engine.api.deployment.CustomActionRequest(
-        name = actionName,
-        processVersion = process.toEngineProcessVersion,
-        user = user.toManagerUser,
-        params = params
-      )
-
+  ): Future[CustomActionResult] = {
     val maybeProcess = processRepository.fetchLatestProcessDetailsForProcessId[CanonicalProcess](id.id)
     maybeProcess.flatMap {
       case Some(process) if process.isFragment =>
-        actionError(
-          CustomActionForbidden(createCustomAction(process), "Invoke custom action on fragment is forbidden.")
-        )
+        actionError(ProcessIllegalAction.fragment(actionName, id))
       case Some(process) if process.isArchived =>
-        actionError(
-          CustomActionForbidden(createCustomAction(process), "Invoke custom action on archived scenario is forbidden.")
-        )
+        actionError(ProcessIllegalAction.archived(actionName, id))
       case Some(process) =>
-        val actionReq = createCustomAction(process)
-        val manager   = dispatcher.deploymentManagerUnsafe(process.processingType)
+        val manager = dispatcher.deploymentManagerUnsafe(process.processingType)
+        val actionReq = CustomActionRequest(
+          name = actionName,
+          processVersion = process.toEngineProcessVersion,
+          user = user.toManagerUser,
+          params = params
+        )
         manager.customActions.find(_.name == actionName) match {
           case Some(customAction) =>
             implicit val freshnessPolicy: DataFreshnessPolicy = DataFreshnessPolicy.Fresh
-            processStateService.getProcessState(process).flatMap { status =>
-              if (customAction.allowedStateStatusNames.contains(status.status.name)) {
-                manager.invokeCustomAction(actionReq, process.json)
-              } else
-                actionError(CustomActionInvalidStatus(actionReq, status.status.name))
-            }
+            processStateService
+              .getProcessState(process)
+              .flatMap { status =>
+                if (customAction.allowedStateStatusNames.contains(status.status.name)) {
+                  manager.invokeCustomAction(actionReq, process.json)
+                } else
+                  actionError(ProcessIllegalAction(actionName, id, status))
+              }
           case None =>
             actionError(CustomActionNonExisting(actionReq))
         }
@@ -73,7 +71,10 @@ class CustomActionInvokerServiceImpl(
     }
   }
 
-  // FIXME: change returning successful to failed..
-  private def actionError(error: CustomActionError) = Future.successful(Left(error))
+  private def actionError(error: Exception with EspError): Future[CustomActionResult] = Future.failed(error)
 
 }
+
+case class CustomActionNonExisting(request: CustomActionRequest)
+    extends Exception(s"""Action "${request.name}" does not exist""")
+    with NotFoundError