+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ManagementResourcesConcurrentSpec.scala
@@ -45,7 +45,7 @@ class ManagementResourcesConcurrentSpec
         secondStatus = status
       }
       val statuses = List(firstStatus, secondStatus)
-      statuses should contain only (StatusCodes.OK, StatusCodes.Conflict)
+      statuses should contain only (StatusCodes.OK, StatusCodes.Forbidden)
       eventually {
         deploymentManager.deploys.asScala.count(_ == ProcessName(processName)) shouldBe 1
       }