+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ManagementResourcesSpec.scala
@@ -18,7 +18,7 @@ import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName, V
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
 import pl.touk.nussknacker.engine.spel.Implicits._
 import pl.touk.nussknacker.restmodel.processdetails._
-import pl.touk.nussknacker.restmodel.{CustomActionRequest, CustomActionResponse}
+import pl.touk.nussknacker.restmodel.CustomActionRequest
 import pl.touk.nussknacker.test.PatientScalaFutures
 import pl.touk.nussknacker.ui.api.helpers.TestFactory._
 import pl.touk.nussknacker.ui.api.helpers._
@@ -78,7 +78,7 @@ class ManagementResourcesSpec
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.DuringDeploy) {
       deployProcess(processName.value) ~> check {
-        status shouldBe StatusCodes.Conflict
+        status shouldBe StatusCodes.Forbidden
       }
     }
   }
@@ -88,7 +88,7 @@ class ManagementResourcesSpec
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Canceled) {
       cancelProcess(processName.value) ~> check {
-        status shouldBe StatusCodes.Conflict
+        status shouldBe StatusCodes.Forbidden
       }
     }
   }
@@ -99,8 +99,8 @@ class ManagementResourcesSpec
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Canceled) {
       deployProcess(processName.value) ~> check {
-        status shouldBe StatusCodes.Conflict
-        responseAs[String] shouldBe ProcessIllegalAction.archived(ProcessActionType.Deploy, processIdWithName).message
+        status shouldBe StatusCodes.Forbidden
+        responseAs[String] shouldBe ProcessIllegalAction.archived(ProcessActionType.Deploy.toString, processIdWithName).message
       }
     }
   }
@@ -110,8 +110,8 @@ class ManagementResourcesSpec
     val processIdWithName = ProcessIdWithName(id, processName)
 
     deployProcess(processName.value) ~> check {
-      status shouldBe StatusCodes.Conflict
-      responseAs[String] shouldBe ProcessIllegalAction.fragment(ProcessActionType.Deploy, processIdWithName).message
+      status shouldBe StatusCodes.Forbidden
+      responseAs[String] shouldBe ProcessIllegalAction.fragment(ProcessActionType.Deploy.toString, processIdWithName).message
     }
   }
 
@@ -120,8 +120,8 @@ class ManagementResourcesSpec
     val processIdWithName = ProcessIdWithName(id, processName)
 
     deployProcess(processName.value) ~> check {
-      status shouldBe StatusCodes.Conflict
-      responseAs[String] shouldBe ProcessIllegalAction.fragment(ProcessActionType.Deploy, processIdWithName).message
+      status shouldBe StatusCodes.Forbidden
+      responseAs[String] shouldBe ProcessIllegalAction.fragment(ProcessActionType.Deploy.toString, processIdWithName).message
     }
   }
 
@@ -416,7 +416,6 @@ class ManagementResourcesSpec
     createEmptyProcess(SampleProcess.processName)
     customAction(SampleProcess.processName, CustomActionRequest("hello")) ~> check {
       status shouldBe StatusCodes.OK
-      responseAs[CustomActionResponse] shouldBe CustomActionResponse(isSuccess = true, msg = "Hi")
     }
   }
 
@@ -424,10 +423,6 @@ class ManagementResourcesSpec
     createEmptyProcess(SampleProcess.processName)
     customAction(SampleProcess.processName, CustomActionRequest("non-existing")) ~> check {
       status shouldBe StatusCodes.NotFound
-      responseAs[CustomActionResponse] shouldBe CustomActionResponse(
-        isSuccess = false,
-        msg = "non-existing is not existing"
-      )
     }
   }
 
@@ -435,10 +430,6 @@ class ManagementResourcesSpec
     createEmptyProcess(SampleProcess.processName)
     customAction(SampleProcess.processName, CustomActionRequest("not-implemented")) ~> check {
       status shouldBe StatusCodes.NotImplemented
-      responseAs[CustomActionResponse] shouldBe CustomActionResponse(
-        isSuccess = false,
-        msg = "not-implemented is not implemented"
-      )
     }
   }
 
@@ -446,10 +437,6 @@ class ManagementResourcesSpec
     createEmptyProcess(SampleProcess.processName)
     customAction(SampleProcess.processName, CustomActionRequest("invalid-status")) ~> check {
       status shouldBe StatusCodes.Forbidden
-      responseAs[CustomActionResponse] shouldBe CustomActionResponse(
-        isSuccess = false,
-        msg = s"Scenario status: NOT_DEPLOYED is not allowed for action invalid-status"
-      )
     }
   }