+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManager.scala
@@ -216,6 +216,6 @@ class PeriodicDeploymentManager private[periodic] (
   override def invokeCustomAction(
       actionRequest: CustomActionRequest,
       canonicalProcess: CanonicalProcess
-  ): Future[Either[CustomActionError, CustomActionResult]] =
+  ): Future[CustomActionResult] =
     customActionsProvider.invokeCustomAction(actionRequest, canonicalProcess)
 }