+++ docs/Changelog.md
@@ -22,6 +22,7 @@
 * [#4707](https://github.com/TouK/nussknacker/pull/4707) Support for `overrideFrontendAuthenticationStrategy` configuration parameter in OIDC security model - works the same as in OAuth2 case.
 * [#4739](https://github.com/TouK/nussknacker/pull/4739) Add configuration parameter for sending additional headers to InfluxDB (`countsSettings.additionalHeaders`)
 * [#4762](https://github.com/TouK/nussknacker/pull/4762) Fix: RegExpParameterValidator, trimming SPeL comprehension
+* [#4687](https://github.com/TouK/nussknacker/pull/4687) Custom action invocation handles errors in the same manner as predefined actions, e.g. cancel.
 
 1.11.3 (11 Sep 2023)
 -------------------------